requirements.txt
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
idna==3.3
mutagen==1.45.1
packaging==21.3
Pillow==9.2.0
pkg_resources==0.0.0
pygame==2.1.2
pyparsing==3.0.9
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
tk==0.1.0
urllib3==1.26.11
wrapt==1.14.1
